factored out types into own file
